+++ CHANGELOG.md
@@ -3,8 +3,15 @@ All notable changes to this project will be documented in this file. This change
 
 ## [Unreleased]
 
+## [0.2.0] - 2025-08-28
+### Added
+- step composition
+- short-circuiting
+- wilful failure
+
 ## 0.1.0 - 2025-08-20
 ### Added
 - project scaffold
 
-[Unreleased]: https://github.com/eureton/workflow/compare/0.1.0...HEAD
+[Unreleased]: https://github.com/eureton/workflow/compare/0.2.0...HEAD
+[0.2.0]: https://github.com/eureton/workflow/compare/0.1.0...0.2.0